Refactoring. Move the post request to createForm module. Add style
